Add files via upload
Showcasing myself reviewing some Nessus logs ( all on practice logs), setting up Splunk alerts and then responding to them whenever something was alerted to them was definitely one of the highlights of my learning.  Whether it was setting up alerts for account password lockouts and thresholds to doing alerts for countries accessing the site that should not be allowed it was all fun and super intriguing!